Fix issue that came up with test case of y=0 in mod operator
with y = 0 causes nan for ref but if sign is negative std::fmod version
preserves sign where gpu mod doesn't which breaks our test.
